RUE-1802: extract the move/borrow/exclusivity state machine
Move VariableMoveState, LoopEdgeStates, CallLoanKind, the expression
exclusivity ledgers, and the move-map join helpers out of sema/context.rs
into a new sema/ownership_state.rs, and gather the eleven ownership
fields AnalysisContext threaded through analysis into one OwnershipState
value embedded as its ownership field. The state transitions - scope
frame save/restore with loop-edge snapshot replay, branch and arm joins,
full-expression ledger nesting, and the loop back-edge recheck fork -
are methods on OwnershipState, unit-tested directly without the engine
(22 migrated VariableMoveState tests plus new coverage for scope
shadowing, loop-edge replay, ledger nesting, the recheck fork, and
branch merges). Consumer sites now reach the state through
ctx.ownership; the consistency pins fencing raw move-state access into
analysis/ownership.rs and control_flow.rs are updated to the new
spelling, and both source inventories list the new module.
