+++ crates/rue-air/src/sema/analysis/builtin_ops.rs
@@ -295,7 +295,7 @@ impl<H: OrdinaryBodyAnalysisHost> OrdinaryBodyEngine<'_, H> {
         F: FnOnce(AirRef, AirRef) -> AirInstData,
     {
         let lhs_result = self.analyze_inst(air, lhs, ctx)?;
-        let reachable_edges_after_lhs = ctx.loop_break_stack.clone();
+        let reachable_edges_after_lhs = ctx.ownership.loop_break_stack.clone();
         let divergence_before_rhs = ctx.divergence_kinds;
         let rhs_result = self.analyze_inst(air, rhs, ctx)?;
         if !lhs_result.continues {
@@ -367,7 +367,7 @@ impl<H: OrdinaryBodyAnalysisHost> OrdinaryBodyEngine<'_, H> {
         );
         let (lhs_result, rhs_result) = if lhs_is_literal {
             let rhs_result = self.analyze_inst_for_projection(air, rhs, ctx)?;
-            let reachable_edges_after_rhs = ctx.loop_break_stack.clone();
+            let reachable_edges_after_rhs = ctx.ownership.loop_break_stack.clone();
             let divergence_before_lhs = ctx.divergence_kinds;
             let expected = (self.is_strbuf(rhs_result.ty) || self.is_str_like(rhs_result.ty))
                 .then_some(rhs_result.ty);
@@ -381,7 +381,7 @@ impl<H: OrdinaryBodyAnalysisHost> OrdinaryBodyEngine<'_, H> {
             (lhs_result, rhs_result)
         } else {
             let lhs_result = self.analyze_inst_for_projection(air, lhs, ctx)?;
-            let reachable_edges_after_lhs = ctx.loop_break_stack.clone();
+            let reachable_edges_after_lhs = ctx.ownership.loop_break_stack.clone();
             let divergence_before_rhs = ctx.divergence_kinds;
             let expected = (self.is_strbuf(lhs_result.ty) || self.is_str_like(lhs_result.ty))
                 .then_some(lhs_result.ty);
