+++ crates/rue-air/src/sema/mod.rs
@@ -46,6 +46,7 @@ mod info;
 mod known_symbols;
 mod ordinary_engine;
 mod output;
+mod ownership_state;
 pub mod provider;
 mod provider_body_host;
 mod provider_module_registry;
