+++ crates/rue-air/src/sema/ordinary_engine.rs
@@ -2301,17 +2301,15 @@ impl<'h, H: OrdinaryBodyAnalysisHost> OrdinaryBodyEngine<'h, H> {
             next_slot: 0,
             loop_depth: 0,
             checked_depth: 0,
-            loop_break_stack: Vec::new(),
             used_locals: AHashSet::new(),
             return_type,
             scope_stack: Vec::new(),
-            moved_scope_stack: Vec::new(),
             comptime_type_scope_stack: Vec::new(),
             resolved_types: &resolved_types,
             resolved_continues: &resolved_continues,
             comptime_selections: &comptime_selections,
             divergence_kinds: DivergenceKinds::NONE,
-            moved_vars: AHashMap::new(),
+            ownership: super::ownership_state::OwnershipState::default(),
             warnings: Vec::new(),
             allow_unused_variables: allow_unused_variable,
             local_string_table: AHashMap::new(),
@@ -2321,19 +2319,11 @@ impl<'h, H: OrdinaryBodyAnalysisHost> OrdinaryBodyEngine<'h, H> {
             comptime_value_vars,
             referenced_functions: AHashSet::new(),
             referenced_methods: std::collections::HashSet::new(),
-            byref_arg_root: None,
-            drop_intrinsic_operand: None,
-            call_loaned_roots: Vec::new(),
-            in_loop_move_recheck: false,
-            iter_borrows: Vec::new(),
             expected_type: None,
             infer_ctx,
             accessor_trailing_yield: None,
             accessor_call_insts: AHashMap::new(),
             accessor_place_refs: AHashMap::new(),
-            expression_loans: Vec::new(),
-            expression_shared_reads: Vec::new(),
-            expression_exclusive_uses: Vec::new(),
             inline_resolved_types: Vec::new(),
             place_aliases: AHashMap::new(),
             try_operand: false,
